refactor(proxy-checker): extract FileCache to separate module
Add new `proxy_checker/FileCache.py` containing the `FileCache` class.
Remove the duplicate `FileCache` implementation from `proxy_checker/ProxyChecker.py`
and import the class from the new module. Export `FileCache` from
`proxy_checker/__init__.py`.
